+++ web/src/App.tsx
@@ -4,59 +4,7 @@ import Game from "./components/game/Game";
 import LandingPage from "./pages/landing/landingPage";
 import JoinPage from "./pages/joinPage/joinPage";
 import "./assets/variables.css";
-
-const avatars = [
-  "bat",
-  "beaver",
-  "bee",
-  "beetle",
-  "boar",
-  "buffalo",
-  "bullfinch",
-  "butterfly",
-  "camel",
-  "cat",
-  "chameleon",
-  "chicken",
-  "clownfish",
-  "cow",
-  "crab",
-  "crocodile",
-  "deer",
-  "dog",
-  "elephant",
-  "flamingo",
-  "fox",
-  "frog",
-  "giraffe",
-  "gorilla",
-  "hedgehog",
-  "hippo",
-  "horse",
-  "ladybug",
-  "lama",
-  "lion",
-  "mouse",
-  "owl",
-  "panda",
-  "parrot",
-  "penguin",
-  "pig",
-  "platypus",
-  "rabbit",
-  "rhino",
-  "shark",
-  "sheep",
-  "sloth",
-  "snake",
-  "spider",
-  "squid",
-  "stingray",
-  "turtle",
-  "walrus",
-  "whale",
-  "zebra",
-];
+import avatars from "./assets/avatars";
 
 // import "./assets";
 const randIndex: () => number = () => {
