+++ web/src/assets/avatars.tsx
@@ -1,0 +1,54 @@
+const avatars = [
+  "bat",
+  "beaver",
+  "bee",
+  "beetle",
+  "boar",
+  "buffalo",
+  "bullfinch",
+  "butterfly",
+  "camel",
+  "cat",
+  "chameleon",
+  "chicken",
+  "clownfish",
+  "cow",
+  "crab",
+  "crocodile",
+  "deer",
+  "dog",
+  "elephant",
+  "flamingo",
+  "fox",
+  "frog",
+  "giraffe",
+  "gorilla",
+  "hedgehog",
+  "hippo",
+  "horse",
+  "ladybug",
+  "lama",
+  "lion",
+  "mouse",
+  "owl",
+  "panda",
+  "parrot",
+  "penguin",
+  "pig",
+  "platypus",
+  "rabbit",
+  "rhino",
+  "shark",
+  "sheep",
+  "sloth",
+  "snake",
+  "spider",
+  "squid",
+  "stingray",
+  "turtle",
+  "walrus",
+  "whale",
+  "zebra",
+];
+
+export default avatars;
