+++ web/src/pages/landing/landingPage.tsx
@@ -6,59 +6,7 @@ import Input from "../../components/input/Input";
 import "./landingPage.css";
 import LandingContainer from "../../containers/landingContainers/landingContainer";
 import LandingBackground from "../../containers/landingBackground/landingBackground";
-
-const avatars = [
-  "bat",
-  "beaver",
-  "bee",
-  "beetle",
-  "boar",
-  "buffalo",
-  "bullfinch",
-  "butterfly",
-  "camel",
-  "cat",
-  "chameleon",
-  "chicken",
-  "clownfish",
-  "cow",
-  "crab",
-  "crocodile",
-  "deer",
-  "dog",
-  "elephant",
-  "flamingo",
-  "fox",
-  "frog",
-  "giraffe",
-  "gorilla",
-  "hedgehog",
-  "hippo",
-  "horse",
-  "ladybug",
-  "lama",
-  "lion",
-  "mouse",
-  "owl",
-  "panda",
-  "parrot",
-  "penguin",
-  "pig",
-  "platypus",
-  "rabbit",
-  "rhino",
-  "shark",
-  "sheep",
-  "sloth",
-  "snake",
-  "spider",
-  "squid",
-  "stingray",
-  "turtle",
-  "walrus",
-  "whale",
-  "zebra",
-];
+import avatars from "../../assets/avatars";
 
 interface LandingProps {
   nickname: string;
